Use "datum" objects to unify program syntax and runtime values
A "datum" is defined by the R5RS standard as anything parsable by (read)
or inside a (quote ...) expression. Since every Scheme expression is also a datum,
we can use unify expressions *and* runtime values under this umbrella.
This lets us implements (quote ...) in the language very easily, and
much of the parsing and serialization code is now shared or very similar.